LayoutTests/platform/wpe/imported/w3c/web-platform-tests/html/browsers/windows/browsing-context-names/choose-_self-002-expected.txt

CONSOLE MESSAGE: QuotaExceededError: The quota has been exceeded.

Harness Error (TIMEOUT), message = null

TIMEOUT choosing _self context should be case-insensitive Test timed out
